# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PCB Inspector AI",
    "description": "Local single-user demo. No trained detector is installed.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/me": {
      "get": {
        "summary": "Identity",
        "operationId": "identity_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Principal"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Ready Ready Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/upload": {
      "post": {
        "summary": "Upload",
        "operationId": "upload_api_v1_inspections_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_inspections_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Submission"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inspections": {
      "get": {
        "summary": "History",
        "operationId": "history_api_v1_inspections_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/{inspection_id}": {
      "get": {
        "summary": "Status",
        "operationId": "status_api_v1_inspections__inspection_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Inspection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/{inspection_id}/results": {
      "get": {
        "summary": "Results",
        "operationId": "results_api_v1_inspections__inspection_id__results_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Inspection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/{inspection_id}/report": {
      "get": {
        "summary": "Download Report",
        "operationId": "download_report_api_v1_inspections__inspection_id__report_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Inspection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/{inspection_id}/image": {
      "get": {
        "summary": "Original Image",
        "operationId": "original_image_api_v1_inspections__inspection_id__image_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Inspection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_api_v1_inspections_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_inspections_upload_post"
      },
      "BoundingBox": {
        "properties": {
          "x1": {
            "type": "number",
            "minimum": 0.0,
            "title": "X1"
          },
          "y1": {
            "type": "number",
            "minimum": 0.0,
            "title": "Y1"
          },
          "x2": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "X2"
          },
          "y2": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Y2"
          }
        },
        "type": "object",
        "required": [
          "x1",
          "y1",
          "x2",
          "y2"
        ],
        "title": "BoundingBox"
      },
      "DefectType": {
        "type": "string",
        "enum": [
          "missing_component",
          "misaligned_component",
          "solder_bridge",
          "damaged_component",
          "board_damage",
          "foreign_object",
          "contamination"
        ],
        "title": "DefectType"
      },
      "Finding": {
        "properties": {
          "defect_type": {
            "$ref": "#/components/schemas/DefectType"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "bbox": {
            "$ref": "#/components/schemas/BoundingBox"
          },
          "severity": {
            "type": "string",
            "enum": [
              "critical",
              "high",
              "medium"
            ],
            "title": "Severity"
          },
          "confidence_band": {
            "type": "string",
            "enum": [
              "high",
              "review"
            ],
            "title": "Confidence Band"
          }
        },
        "type": "object",
        "required": [
          "defect_type",
          "confidence",
          "bbox",
          "severity",
          "confidence_band"
        ],
        "title": "Finding"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "History": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InspectionSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "limit",
          "offset"
        ],
        "title": "History"
      },
      "InspectionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "overall_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "completed_at",
          "width",
          "height",
          "attempts",
          "model_version",
          "error_code",
          "overall_result"
        ],
        "title": "InspectionSummary"
      },
      "Principal": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "auth_mode": {
            "type": "string",
            "title": "Auth Mode"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "subject",
          "owner_id",
          "auth_mode"
        ],
        "title": "Principal"
      },
      "Report": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "inspection_id": {
            "type": "string",
            "title": "Inspection Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "overall_result": {
            "type": "string",
            "enum": [
              "NOT_EVALUATED",
              "REVIEW_REQUIRED",
              "NO_VISIBLE_DEFECTS_DETECTED"
            ],
            "title": "Overall Result"
          },
          "inference_time_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Inference Time Ms"
          },
          "image_width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Image Width"
          },
          "image_height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Image Height"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Detections"
          },
          "ignored_detection_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Ignored Detection Count"
          },
          "decision_policy_version": {
            "type": "string",
            "title": "Decision Policy Version",
            "default": "provisional-1"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "inspection_id",
          "model_version",
          "is_demo",
          "overall_result",
          "inference_time_ms",
          "image_width",
          "image_height",
          "detections",
          "ignored_detection_count",
          "limitations"
        ],
        "title": "Report"
      },
      "Status": {
        "type": "string",
        "enum": [
          "QUEUED",
          "PROCESSING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "Status"
      },
      "Submission": {
        "properties": {
          "inspection_id": {
            "type": "string",
            "title": "Inspection Id"
          },
          "status": {
            "type": "string",
            "const": "QUEUED",
            "title": "Status",
            "default": "QUEUED"
          }
        },
        "type": "object",
        "required": [
          "inspection_id"
        ],
        "title": "Submission"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Cognito user-pool access token (RS256)",
        "scheme": "bearer"
      }
    }
  }
}